requirements.txt
python-a2s
discord.py
python-dotenv